% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niffty.R
\name{ifttt_api_key}
\alias{ifttt_api_key}
\title{Get or set IFTTT_KEY value}
\usage{
ifttt_api_key(force = FALSE)
}
\arguments{
\item{force}{force setting a new IFTTT API key for the current environment?}
}
\value{
atomic character vector containing the IFTTT API key
}
\description{
The API wrapper functions in this package all rely on an IFTTT API
key residing in the environment variable \code{IFTTT_API_KEY}. The
easiest way to accomplish this is to set it in the `\code{.Renviron}` file in your
home directory.
}

