% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niffty.R
\name{maker}
\alias{maker}
\title{Issue IFTTT maker channel POST event}
\usage{
maker(event = NULL, value1 = NULL, value2 = NULL, value3 = NULL)
}
\arguments{
\item{event}{event name to trigger in IFTTT recipe}

\item{value1}{event parameter (optional)}

\item{value2}{event parameter (optional)}

\item{value3}{event parameter (optional)}
}
\description{
Issue IFTTT maker channel POST event
}

