% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nifffty-package.r
\docType{package}
\name{nifffty}
\alias{nifffty}
\alias{nifffty-package}
\title{A package to post events to IFTTT 'Maker' channels}
\description{
A package to post events to IFTTT 'Maker' channels
}
\author{
Bob Rudis (bob@rud.is)
}
