% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/app.R
\name{receiver}
\alias{receiver}
\title{Listen and react to IFTTT Maker web calls}
\usage{
receiver(listen = "0.0.0.0", port = 9999, handler = NULL)
}
\arguments{
\item{port}{TCP port to listen on. Defaults to \code{9999}}

\item{handler}{R function to call with the results of the Maker web call}

\item{listem}{IP address to listen on. \code{127.0.0.1} for localhost,
\code{0.0.0.0} (default) for all interfaces.}
}
\description{
Listen and react to IFTTT Maker web calls
}

