% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niffty.R
\name{maker}
\alias{maker}
\title{Issue IFTTT maker channel POST event}
\usage{
maker(event = NULL, value1 = NULL, value2 = NULL, value3 = NULL)
}
\arguments{
\item{event}{event name to trigger in IFTTT recipe}

\item{value1}{event parameter (optional)}

\item{value2}{event parameter (optional)}

\item{value3}{event parameter (optional)}
}
\description{
You must have an IFTTT account and an event trigget setup with an
IFTTT Maker channel. See
\href{http://bconnelly.net/2015/06/connecting-r-to-everything-with-ifttt/}{this post} for
more information and a full example.
}
\seealso{
\url{http://bconnelly.net/2015/06/connecting-r-to-everything-with-ifttt/}
}

